spacetime/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
yfinance>=0.2.37
reportlab>=4.1.0
matplotlib>=3.8.0
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.12.0
sqlalchemy>=2.0.0
alpaca-py>=0.20.0
pyarrow>=15.0.0
plotly>=5.20.0
pydantic>=2.6.0
